Split transcoder-specific code into its own file.
